+++ noodles-fasta/src/fai/io.rs
@@ -1,0 +1,5 @@
+//! FAI I/O.
+
+mod reader;
+
+pub use self::reader::Reader;